+++ src/color.h
@@ -1,8 +1,6 @@
 
-uint32_t color_char_to_packed_col(const char *color);
-uint32_t color_sexp_to_packed_col(SEXP color_);
-SEXP packed_col_to_CHARSXP_color(uint32_t *packed_col);
-uint32_t *colors_to_packed_cols(SEXP colors_, int N, bool *do_free);
+uint32_t single_rcolor_to_int(SEXP color_);
+uint32_t *multi_rcolors_to_ints(SEXP colors_, int N, bool *do_free);
 
 bool is_transparent(uint32_t color);
 bool is_opaque(uint32_t color);