+++ src/hash-color.h
@@ -1,1 +1,0 @@
-uint32_t rcolor_to_int(const char *col);